Get rid of outline
